+++ web/list/mail/index.php
@@ -9,62 +9,30 @@ include $_SERVER["DOCUMENT_ROOT"] . "/inc/main.php";
 
 // Data & Render page
 if (empty($_GET["domain"])) {
-	exec(HESTIA_CMD . "h-list-mail-domains $user json", $output, $return_var);
-	$data = json_decode(implode("", $output), true);
+	$data = cli_json("h-list-mail-domains $user json");
 	if ($_SESSION["userSortOrder"] == "name") {
 		ksort($data);
 	} else {
 		$data = array_reverse($data, true);
 	}
-	unset($output);
 
 	render_page($user, $TAB, "list_mail");
 } elseif (!empty($_GET["dns"])) {
-	exec(
-		HESTIA_CMD . "h-list-mail-domain " . $user . " " . quoteshellarg($_GET["domain"]) . " json",
-		$output,
-		$return_var,
+	$data = array_reverse(cli_json("h-list-mail-domain " . $user . " " . quoteshellarg($_GET["domain"]) . " json"), true);
+	$ips = array_reverse(cli_json("h-list-user-ips " . $user . " json"), true);
+	$dkim = array_reverse(
+		cli_json("h-list-mail-domain-dkim-dns " . $user . " " . quoteshellarg($_GET["domain"]) . " json"),
+		true,
 	);
-	$data = json_decode(implode("", $output), true);
-	$data = array_reverse($data, true);
-	unset($output);
-	exec(HESTIA_CMD . "h-list-user-ips " . $user . " json", $output, $return_var);
-	$ips = json_decode(implode("", $output), true);
-	$ips = array_reverse($ips, true);
-	unset($output);
-	exec(
-		HESTIA_CMD .
-			"h-list-mail-domain-dkim-dns " .
-			$user .
-			" " .
-			quoteshellarg($_GET["domain"]) .
-			" json",
-		$output,
-		$return_var,
-	);
-	$dkim = json_decode(implode("", $output), true);
-	$dkim = array_reverse($dkim, true);
-	unset($output);
 
 	render_page($user, $TAB, "list_mail_dns");
 } else {
-	exec(
-		HESTIA_CMD .
-			"h-list-mail-accounts " .
-			$user .
-			" " .
-			quoteshellarg($_GET["domain"]) .
-			" json",
-		$output,
-		$return_var,
-	);
-	$data = json_decode(implode("", $output), true);
+	$data = cli_json("h-list-mail-accounts " . $user . " " . quoteshellarg($_GET["domain"]) . " json");
 	if ($_SESSION["userSortOrder"] == "name") {
 		ksort($data);
 	} else {
 		$data = array_reverse($data, true);
 	}
-	unset($output);
 
 	render_page($user, $TAB, "list_mail_acc");
 }